[ refactor ] Modulize `syntax::abs`
